requirements.txt
asgiref==3.6.0
backports.zoneinfo==0.2.1;python_version<"3.9"
black==22.12.0
cffi==1.15.1
click==8.1.3
cryptography==39.0.0
Django==4.1.5
django-cors-headers==3.13.0
djangorestframework==3.14.0
isort==5.11.4
mypy-extensions==0.4.3
pathspec==0.10.3
platformdirs==2.6.2
psycopg2==2.9.5
psycopg2-binary==2.9.5
pycparser==2.21
PyJWT==2.6.0
pytz==2022.7.1
sqlparse==0.4.3
tomli==2.0.1
typing-extensions==4.4.0
